requirements.txt
ocrmypdf
pypdf
pikepdf
python-dotenv
llama-parse
unstructured
unstructured_inference
pdf2image
unstructured_pytesseract
pdfplumber
pandas
pypdf